Adding function and class checks while also adding phan suppressions that are related

+++ projects/packages/classic-theme-helper/src/social-links.php
@@ -17,6 +17,8 @@
 
 /**
  * Init Social_Links if the theme declares support.
+ *
+ * @suppress PhanNoopNew
  */
 function jetpack_theme_supports_social_links() {
 	if ( ! wp_is_block_theme() && current_theme_supports( 'social-links' ) && function_exists( 'publicize_init' ) ) {
@@ -76,7 +78,7 @@ if ( ! class_exists( 'Social_Links' ) ) {
 			}
 
 			$this->theme_supported_services = $theme_support[0];
-			$this->links                    = Jetpack_Options::get_option( 'social_links', array() );
+			$this->links                    = class_exists( \Jetpack_Options::class ) ? Jetpack_Options::get_option( 'social_links', array() ) : '';
 
 			$this->admin_setup();
 
@@ -101,7 +103,8 @@ if ( ! class_exists( 'Social_Links' ) ) {
 				return;
 			}
 
-			$this->publicize    = publicize_init();
+			// @phan-suppress-next-line PhanUndeclaredFunction -- Function checked with function_exists - see https://github.com/phan/phan/issues/1204.
+			$this->publicize    = function_exists( 'publicize_init' ) ? publicize_init() : null;
 			$publicize_services = $this->publicize->get_services( 'connected' );
 			$this->services     = array_intersect( array_keys( $publicize_services ), $this->theme_supported_services );
 
@@ -122,7 +125,9 @@ if ( ! class_exists( 'Social_Links' ) ) {
 
 			if ( $active_links !== $this->links ) {
 				$this->links = $active_links;
-				Jetpack_Options::update_option( 'social_links', $active_links );
+				if ( class_exists( \Jetpack_Options::class ) ) {
+					Jetpack_Options::update_option( 'social_links', $active_links );
+				}
 			}
 		}
 
@@ -140,31 +145,33 @@ if ( ! class_exists( 'Social_Links' ) ) {
 				)
 			);
 
-			foreach ( array_keys( $this->publicize->get_services( 'all' ) ) as $service ) {
-				$choices = $this->get_customize_select( $service );
+			if ( class_exists( \Publicize::class ) ) {
+				foreach ( array_keys( $this->publicize->get_services( 'all' ) ) as $service ) {
+					$choices = $this->get_customize_select( $service );
 
-				if ( empty( $choices ) ) {
-					continue;
+					if ( empty( $choices ) ) {
+						continue;
+					}
+
+					$wp_customize->add_setting(
+						"jetpack_options[social_links][$service]",
+						array(
+							'type'    => 'option',
+							'default' => '',
+						)
+					);
+
+					$wp_customize->add_control(
+						"jetpack-$service",
+						array(
+							'label'    => $this->publicize->get_service_label( $service ),
+							'section'  => 'jetpack_social_links',
+							'settings' => "jetpack_options[social_links][$service]",
+							'type'     => 'select',
+							'choices'  => $choices,
+						)
+					);
 				}
-
-				$wp_customize->add_setting(
-					"jetpack_options[social_links][$service]",
-					array(
-						'type'    => 'option',
-						'default' => '',
-					)
-				);
-
-				$wp_customize->add_control(
-					"jetpack-$service",
-					array(
-						'label'    => $this->publicize->get_service_label( $service ),
-						'section'  => 'jetpack_social_links',
-						'settings' => "jetpack_options[social_links][$service]",
-						'type'     => 'select',
-						'choices'  => $choices,
-					)
-				);
 			}
 		}
 
@@ -233,16 +240,18 @@ if ( ! class_exists( 'Social_Links' ) ) {
 				$choices[ $this->links[ $service ] ] = $this->links[ $service ];
 			}
 
-			$connected_services = $this->publicize->get_services( 'connected' );
-			if ( isset( $connected_services[ $service ] ) ) {
-				foreach ( $connected_services[ $service ] as $c ) {
-					$profile_link = $this->publicize->get_profile_link( $service, $c );
+			if ( class_exists( \Publicize::class ) ) {
+				$connected_services = $this->publicize->get_services( 'connected' );
+				if ( isset( $connected_services[ $service ] ) ) {
+					foreach ( $connected_services[ $service ] as $c ) {
+						$profile_link = $this->publicize->get_profile_link( $service, $c );
 
-					if ( false === $profile_link ) {
-						continue;
+						if ( false === $profile_link ) {
+							continue;
+						}
+
+						$choices[ $profile_link ] = $this->publicize->get_display_name( $service, $c );
 					}
-
-					$choices[ $profile_link ] = $this->publicize->get_display_name( $service, $c );
 				}
 			}
